- Updated Build scripts to use VS2015 as default compiler for building the framework and SDK - Redirected VB and C$ compilers to use the pre-Roslyn compiler as Roslyn generates code assuming a full V4.0 runtime (including Generics) which makes compiling for NETMF difficult. - silenced multiple bening warnings detected by newer VS2015 C++ compiler. - silenced a number of warnings inC# code and tests

+++ DeviceCode/Drivers/LargeBuffer/Test/LargeBuffer_test.cpp
@@ -15,7 +15,9 @@ void LargeBuffer_PostEvent( UINT16 marshalId )
 
 void LargeBufferTest_Completion(void* arg)
 {
-    LargeBuffer_PostEvent( (UINT16)arg );
+    // Double cast to silence compiler warning about shortening pointer type
+    // Ideally this would have used uintptr_t but that wasn't widely supported at the time the API was defined
+    LargeBuffer_PostEvent( (UINT16)(UINT32)arg );
 }
 
 static UINT8 s_Data[800*600*2];
